+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-05-10 18:29:25
-// Commit: chore(test): update dependencies and cleanup unused imports in test
+// Updated at: 2026-05-10 18:52:19
+// Commit: test(test): add unit tests for test business logic
 
 class testComponentState {
-  final String commitRef = '1021';
-  final String lastMessage = 'chore(test): update dependencies and cleanup unused imports in test';
-  final DateTime timestamp = DateTime.parse('2026-05-10T18:29:25+05:00');
+  final String commitRef = '1023';
+  final String lastMessage = 'test(test): add unit tests for test business logic';
+  final DateTime timestamp = DateTime.parse('2026-05-10T18:52:19+05:00');
 }